requirements.txt
rfdetr
Pillow
